Create unfettering.txt
I was reminded earlier today, while talking with Charlie Connolly, of a sculpture park in Ireland called Victoria's Way. I lamented that I had not recommended it to him before he visited the country, but recommended it to him nonetheless in hindsight. There were multiple sculptures in which the child must distinguish one's self from the mother, or the creator, in order to become a man. Here, I illustrate the wanderings of an ignorant child in danger in contrast with the knowledge of the knowing mother. The mother must let go of the child for it to pursue meaningful dreams, while the child must be free of restriction and cloistering from the parent in order to venture to any legitimate discovery. Balance must be stricken.