Fix person-number and adds further recursion
I realized that I fell off of third person usage, so I came back straight away to fix that. Further, by slightly transforming the word entirely to entire, meaning the same symbolic object of a united whole, the poem could fall into a beautiful recursive pattern. AABCBC in syllabic count, but without changing the symbolic meaning of the last line. The size of the meaning-space of each line is compilable to the shape of a square-edged triangle with the two edges facing downward and to the right: ◸. DL 7//18 Earth Now 10:45pm