Create broken-kinship.txt
This one came to me strongly in an image and sound of great loss bottled within. To let out the pain in company is the only way to usher consolation to the heart of a degree strong enough to dampen the flames of tragedy. Let yourself wail from the pain inside to unify you with the family you find dwelling outside. Love restores.